REFACTOR: 프로젝트 디테일 페이지 레이아웃과 단일화
